server/requirements.txt
aiofiles==25.1.0
alembic==1.18.2
authlib==1.6.6
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
click==8.3.1
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pycrdt==0.12.45
pycrdt-store==0.1.3
pycrdt-websocket==0.16.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
SQLAlchemy==2.0.46
sqlite-anyio==0.2.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
websockets==16.0
